+++ packages/react-querybuilder/src/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useAsyncOptionList';
 export * from './useControlledOrUncontrolled';
 export * from './useDeprecatedProps';
 export * from './useFields';
@@ -9,4 +10,3 @@ export * from './usePrevious';
 export * from './useReactDndWarning';
 export * from './useSelectElementChangeHandler';
 export * from './useStopEventPropagation';
-export * from './useValueSelectorAsync';